Improve ADO.Net provider tests for QA
