+++ DubUrl.QA/Sqlite/AdoProviderSqlite.cs
@@ -1,25 +1,19 @@
 using NUnit.Framework;
-using System.Diagnostics;
-using System.Data;
-using System.Data.Common;
-using DubUrl.Querying;
-using DubUrl.Querying.Reading;
-using DubUrl.Registering;
-using DubUrl.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using System.Reflection;
 
-namespace DubUrl.QA.Mysql
+namespace DubUrl.QA.Sqlite
 {
-    [Category("MySQL")]
+    [Category("Sqlite")]
     [Category("AdoProvider")]
-    public class AdoProvider : BaseAdoProvider
+    public class AdoProviderSqlite : BaseAdoProvider
     {
+        protected string CurrentDirectory
+        {
+            get => Path.GetDirectoryName(GetType().Assembly.Location) ?? throw new NotImplementedException();
+        }
+
         public override string ConnectionString
         {
-            get => $"mysql://root:Password12!@localhost/DubUrl";
+            get => $"sqlite:///{CurrentDirectory}\\Customer.db";
         }
 
         [Test]
@@ -36,7 +30,7 @@ namespace DubUrl.QA.Mysql
 
         [Test]
         public override void QueryCustomerWithPositionalParameter()
-            => QueryCustomerWithPositionalParameter("select FullName from Customer where CustomerId=?");
+            => Assert.Ignore("select FullName from Customer where CustomerId=($1)");
 
         [Test]
         public override void QueryCustomerWithDapper()
